+++ snackbezorgd-backend/snackbezorgd/snackbezorgd_app/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from .models import product, order  # Make sure to import the 'Order' model
 
 class SnackBezorgdAdmin(admin.ModelAdmin):
-    list_display = ('title', 'description', 'price', 'active')
+    list_display = ('product_number', 'title', 'description', 'price', 'active')
 
 class OrderAdmin(admin.ModelAdmin):  # Create a new admin class for the 'Order' model if needed
     list_display = ('order_number', 'first_name', 'last_name', 'email', 'time', 'paid', 'finished', 'total')
