+++ snackbezorgd-backend/snackbezorgd/snackbezorgd_app/views.py
@@ -1,16 +1,16 @@
 from django.shortcuts import render
 from rest_framework import viewsets
-from .serializers import SnackBezorgdSerializer
+from .serializers import *
 from .models import *
 
 # Create your views here.
 
 class SnackBezorgdView(viewsets.ModelViewSet):
-    serializer_class = SnackBezorgdSerializer
+    serializer_class = ProductSerializer
     queryset = product.objects.all()
 
 class OrderAdmin(viewsets.ModelViewSet):
-    serializer_class = SnackBezorgdSerializer
+    serializer_class = OrderSerializer
     queryset = order.objects.all()
