+++ snackbezorgd-frontend/snackbezorgd/src/pages/home.js
@@ -71,6 +71,7 @@ class Home extends Component {
     };
   }
 
+  
   render() {
     return (
       <Stack>
